Progress
I DON'T understand the forms anymore, so awk (what are they supposed to
have).
Did: commented out functions that aren't needed because the drop down is
gone, made separate forms "work" (both go to matt's page instead of
separate pages), clicking outside the form closes the dialog.

To do: getting to matt's page properly, and getting edit search to work.

Notes: the form is buggy if you don't enter all the fields, originally I
had a message pop up, but matt removed the box, so it's gone. I'm going
to put it back (yes?). Lots of unnecessary code, but it's too late to
fix it now.